added report repository updated test data.config reference to reference the one in the .data directory

+++ Code/EnvironmentalApp/EnvironmentalApp.Data/SQLServer/Repositories/ChilledWater_Campus_DailyTotals_SQL_Repository.cs
@@ -18,7 +18,7 @@ namespace EnvironmentalApp.Data.SQLServer.Repositories
                 {
 
                     var dailyTotals = new CW_DailyTotals_Campus();
-                    var readings = (List<decimal>)entityList.Select(x => x.Reading).ToList();
+                    var readings = (List<float>)entityList.Select(x => x.Reading).ToList();
 
                     dailyTotals.Id = Guid.NewGuid();
                     dailyTotals.ReadingDateTime = DateTime.Now;
