+++ Code/EnvironmentalApp/EnvironmentalApp.Data/SQLServer/Repositories/Electric_DailyTotals_SQL_Repository.cs
@@ -18,7 +18,7 @@ namespace EnvironmentalApp.Data.SQLServer.Repositories
                 {
                  
                     var dailyTotals = new ElectricDailyTotals();
-                    var readings = (List<decimal>)entityList.Select(x => x.Reading).ToList();
+                    var readings = (List<float>)entityList.Select(x => x.Reading).ToList();
 
                     dailyTotals.Id = Guid.NewGuid();
                     dailyTotals.ReadingDateTime = DateTime.Now;
