+++ Code/EnvironmentalApp/EnvironmentalApp.Data/SQLServer/Repositories/Steam_DailyTotals_SQL_Repository.cs
@@ -20,7 +20,7 @@ namespace EnvironmentalApp.Data.SQLServer.Repositories
                     //var SteamTotals = new List<SteamDailyTotals>();
 
                     var dailyTotals = new SteamDailyTotals();
-                    var readings = (List<decimal>)entityList.Select(x => x.Reading).ToList();
+                    var readings = (List<float>)entityList.Select(x => x.Reading).ToList();
 
                     dailyTotals.Id = Guid.NewGuid();
                     dailyTotals.ReadingDateTime = DateTime.Now;
